Move remote.fetcher
remote/image.go -> remote/descriptor.go
